Great relayoutening
Tries to drag everything to be way more like the good ship faff and
brings back the tradition of the snoot being the armory.

The wiring is still a complete mess but it now has significantly more
space to be a mess in, so I can overhaul that next.

Not entirely sure what to do with the excess crew space. I think the
batteries should go in there but it'd be very odd for them to be in
front of it.

O2 should also probably be where cargo is